fix: Move test file into proper location
