+++ tests/client/cli/test_cli_commands.py
@@ -1,22 +1,9 @@
 import re
 
 import pytest
-from aioresponses import aioresponses
-from click.testing import CliRunner
 
-from ai.backend.cli.loader import load_entry_points
 from ai.backend.cli.types import ExitCode
 from ai.backend.client.config import get_config, set_config
-
-
-@pytest.fixture(scope="module")
-def runner():
-    return CliRunner()
-
-
-@pytest.fixture(scope="module")
-def cli_entrypoint():
-    return load_entry_points(allowlist={"ai.backend.client.cli"})
 
 
 @pytest.mark.parametrize("help_arg", ["-h", "--help"])
@@ -85,49 +72,3 @@ def test_run_file_or_code_required(
     result = runner.invoke(cli_entrypoint, ["run", "python"])
     assert result.exit_code == ExitCode.INVALID_ARGUMENT
     assert "provide the command-line code snippet" in result.output
-
-
-@pytest.mark.parametrize(
-    "test_case",
-    [
-        {
-            "session_id_or_name": "00000000-0000-0000-0000-000000000000",
-            "new_session_name": "new-name",
-            "expected_exit_code": ExitCode.OK,
-        },
-        {
-            "session_id_or_name": "mock-session-name",
-            "new_session_name": "new-name",
-            "expected_exit_code": ExitCode.OK,
-        },
-    ],
-    ids=["Use session command by uuid", "Use session command by session name"],
-)
-def test_session_command(
-    test_case, runner, cli_entrypoint, monkeypatch, example_keypair, unused_tcp_port_factory
-):
-    """
-    Test whether the Session CLI commands work correctly when either session_id or session_name is provided as argument.
-    """
-
-    api_port = unused_tcp_port_factory()
-    api_url = "http://127.0.0.1:{}".format(api_port)
-
-    set_config(None)
-    monkeypatch.setenv("BACKEND_ACCESS_KEY", example_keypair[0])
-    monkeypatch.setenv("BACKEND_SECRET_KEY", example_keypair[1])
-    monkeypatch.setenv("BACKEND_ENDPOINT", api_url)
-    monkeypatch.setenv("BACKEND_ENDPOINT_TYPE", "api")
-
-    with aioresponses() as mocked:
-        session_id_or_name = test_case["session_id_or_name"]
-        new_session_name = test_case["new_session_name"]
-
-        mocked.post(
-            f"{api_url}/session/{session_id_or_name}/rename?name={new_session_name}", status=204
-        )
-
-        result = runner.invoke(
-            cli_entrypoint, args=["session", "rename", session_id_or_name, new_session_name]
-        )
-        assert result.exit_code == test_case["expected_exit_code"]
